Fix for https://github.com/coralproject/pillar/issues/21. Consolidated config stuff into config package.
