+++ server/config/config.go
@@ -1,25 +1,50 @@
-package log
+package config
 
 import (
-	"fmt"
-	"log"
 	"os"
+	"log"
 	"strings"
-	"github.com/coralproject/pillar/server/cfg"
+	"fmt"
 )
+
+//Context information for pillar server - Read-Only
+type Context struct {
+	Home      string `json:"home" bson:"home"`
+	MongoURL  string `json:"mongo_url" bson:"mongo_url"` //mongodb://localhost:27017/coral
+}
+var context Context
+
+func GetContext() Context {
+	return context
+}
 
 //Single logger for the application
 var (
 	Logger *log.Logger
 )
 
+//export PILLAR_HOME=path to pillar home
 func init() {
-	logFile := getLogFile()
-	fmt.Printf("Pillar log file: %s\n\n", logFile)
+	home := os.Getenv("PILLAR_HOME")
+	if home == "" {
+		log.Fatal("Error initializing Server: PILLAR_HOME not found.")
+	}
+	//set pillar home
+	context.Home = home
+
+	url := os.Getenv("MONGODB_URL")
+	if home == "" {
+		log.Fatal("Error initializing Server: MONGODB_URL not found.")
+	}
+	context.MongoURL = url
+
+	logFile := getLogFile(home)
 	file, err := os.OpenFile(logFile, os.O_WRONLY|os.O_CREATE|os.O_APPEND, 0666)
 	if err != nil {
 		log.Fatalf("Error opening log file [%s], %s", logFile, err)
 	}
+	fmt.Printf("Pillar log file: %s\n\n", logFile)
+
 
 	Logger = log.New(file, "Pillar: ", log.LstdFlags|log.Llongfile|log.Ldate|log.Ltime)
 }
@@ -40,8 +65,7 @@ func exists(path string) bool {
 	return false
 }
 
-func getLogFile() string {
-	pillarHome := cfg.GetContext().Home
+func getLogFile(pillarHome string) string {
 	if pillarHome == "" {
 		pillarHome = strings.TrimSuffix(os.TempDir(), "/") + "/pillar"
 	}
@@ -53,3 +77,4 @@ func getLogFile() string {
 
 	return logPath + "/pillar.log"
 }
+
